+++ src/Ui/Bundle/AdminBundle/Form/Type/Type/TypeUpdateType.php
@@ -135,6 +135,9 @@ class TypeUpdateType extends AbstractType
             ->add('mustEvaluateMeeting', CheckboxType::class, [
                 'required' => false,
             ])
+            ->add('canSubmitValidation', CheckboxType::class, [
+                'required' => false,
+            ])
         ;
     }